+++ www/OSBarcode.js
@@ -1,5 +1,17 @@
 var exec = require('cordova/exec');
 
-exports.coolMethod = function (arg0, success, error) {
-    exec(success, error, 'OSBarcode', 'coolMethod', [arg0]);
-};
+exports.scanBarcode = function (options, successCallback, errorCallback) {
+    options = options || {};
+
+    let scanInstructions = options.scanInstructions;
+    let cameraDirection = options.cameraDirection;
+    let scanOrientation = options.scanOrientation;
+    let scanButton = !!options.scanButton;
+    let scanText = options.scanText;
+    let hint = options.hint;
+    let androidScanningLibrary = options.androidScanningLibrary;
+
+    let args = [{scanInstructions, cameraDirection, scanOrientation, scanButton, scanText, hint, androidScanningLibrary}];
+
+    exec(successCallback, errorCallback, 'OSBarcode', 'scanBarcode', args);
+}